+++ R/arguments.R
@@ -2,7 +2,9 @@
 #' @name common_arguments
 #' @param data [data.frame] from which the descriptive statistics are
 #'   calculated.
-#' @param id_var [character] The name of the id variable.
+#' @param id_var [character] The name of the id variable. Defaults to `
+#'   getOption("sdc.id_var")` so that you can provide `options(sdc.id_var =
+#'   "my_id_var")` at the top of your script.
 #' @param val_var [character] vector of value variables on which descriptive
 #'   statistics are computed.
 #' @param by Grouping variables (or expression) as in [data.table]'s `by`.