+++ R/utils.R
@@ -1,28 +1,22 @@
 # pure technical settings
 .datatable.aware <- TRUE
 
-#' @importFrom checkmate assert_data_frame assert_string assert check_class
-#' @importFrom methods hasArg
+#' @importFrom checkmate assert_data_frame assert_string assert_subset
 check_args <- function(
   data, id_var, val_var = NULL, by = NULL, zero_as_NA = NULL
 ) {
   checkmate::assert_data_frame(data)
+  col_names <- names(data)
+
   checkmate::assert_string(id_var)
-  if (!is.null(val_var)) {
-    checkmate::assert_string(val_var)
-  }
-  if (methods::hasArg(by)) {
-    checkmate::assert(
-      combine = "or",
-      .var.name = "by",
-      checkmate::check_class(substitute(by), "name"),
-      checkmate::check_class(substitute(by), "call"),
-      checkmate::check_class(substitute(by), "character")
-    )
-  }
-  if (!is.null(zero_as_NA)) {
-    checkmate::assert_logical(zero_as_NA)
-  }
+  checkmate::assert_subset(id_var, choices = names(data))
+
+  checkmate::assert_string(val_var, null.ok = TRUE)
+
+  checkmate::assert_character(by, any.missing = FALSE, null.ok = TRUE)
+  checkmate::assert_subset(by, choices = names(data))
+
+  checkmate::assert_logical(zero_as_NA, len = 1L, null.ok = TRUE)
 }